Add some casts in vpi to remove warnings.
The Cygwin compiler is a bit picky. This patch adds some casts
to remove compilation warnings. In the past I have had warnings
off because of problems with the STL, but for C directories
like this it makes sense to enable the warnings. It also does
not recognize that an assert(0) or assert(false) ends a routine
so it complains about no return at end of function or variables
not being defined.
(cherry picked from commit f31a4e6d20cb03a87785ae72d994bca42a87ca29)

+++ vpi/sys_scanf.c
@@ -650,7 +650,7 @@ static PLI_INT32 sys_fscanf_calltf(PLI_BYTE8*name)
 	    vpi_printf("ERROR: %s:%d: ", vpi_get_str(vpiFile, callh),
 	               (int)vpi_get(vpiLineNo, callh));
 	    vpi_printf("invalid file descriptor (0x%x) given to %s.\n",
-	               val.value.integer, name);
+	               (int)val.value.integer, name);
 	    errno = EBADF;
 	    val.format = vpiIntVal;
 	    val.value.integer = EOF;
